+++ src/otx/engine/adaptive_bs/adaptive_bs_api.py
@@ -1,6 +1,6 @@
 """Algorithm to find a proper batch size which is fit to current GPU device for tasks using mmcv."""
 
-# Copyright (C) 2023 Intel Corporation
+# Copyright (C) 2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
 from __future__ import annotations
